1st Sample of Login Id and Password done.
I want everyone to test.
Download this folder and unzip it. After which open the CMD in this folder and call it a cmd "python app.py" to open the web portal click on the link "http://127.0.0.1:5000" after which on the web portal you will see a page says: "Welcome to the Flask Authentication App"

go to address bar and add login in front of ip as below:
http://127.0.0.1:5000/login

Then click register. Enter your id and password for registration. you will be able to get into login page and login with your user id and password. That id and password will be stored in our SQL database.